% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_Ht.R
\name{corr_ht}
\alias{corr_ht}
\title{Corrected Height calculation}
\usage{
corr_ht(z_score = NULL, z_ht = NULL)
}
\arguments{
\item{z_score}{Taregt height z-score}

\item{z_ht}{Actual height z-score}
}
\description{
This function calculate the corrected height
}
\details{
(Corrected Height SDS = Ht SDS - Target Height SDS)
}
