% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetHt.R
\name{TargetHt}
\alias{TargetHt}
\title{Target height calculation}
\usage{
TargetHt(HtFather = NULL, HtMother = NULL, gender = c("F", "M"))
}
\arguments{
\item{HtFather}{Father's height in cm}

\item{HtMother}{Mother's height in cm}

\item{gender}{Gender of the child F: Female, M: Male}
}
\description{
this function aims to calculate the target height according to gender
}
