% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_ht.R
\name{z_ht}
\alias{z_ht}
\title{z-score of actual height}
\usage{
z_ht(ht, mu, std)
}
\arguments{
\item{ht}{actual height}

\item{mu}{median height value}

\item{std}{Standerd deviation of height}
}
\description{
Calculate z-score of actual height
}
\concept{z score}
