% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Target Z-score calculation}
\usage{
z_score(HtPot = TargetHt, mHt = NULL, sdHt = NULL, L = NULL)
}
\arguments{
\item{HtPot}{Heigh potential which is equal to TargetHt}

\item{mHt}{Median height}

\item{sdHt}{standerd deviation of height}

\item{L}{box-cox transformation}
}
\description{
This funcion aim to calculate the target z-score for Height potential
}
