% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_to_percent.R
\name{z_to_percent}
\alias{z_to_percent}
\title{Z score to percentile}
\usage{
z_to_percent(z_score = NULL)
}
\arguments{
\item{z_score}{z-score calculated from z_score function}
}
\description{
This function will match the z-score to percentile from z-score percentile table
}
